Deprecate tf2_eigen.h
Move it to tf2_eigen.hpp instead.
